Split between legacy and new customizable header
