+++ react/index.js
@@ -1,140 +1,23 @@
-import React, { Fragment, useEffect, useLayoutEffect, useState } from 'react'
-import classNames from 'classnames'
-import PropTypes from 'prop-types'
-import { ExtensionPoint, useRuntime } from 'vtex.render-runtime'
-import TopMenu from './components/TopMenu'
-import Spacer from './components/Helpers/Spacer'
-import { logo, collapsible, icons, searchBar, login } from './defaults'
+import React from 'react'
+import { useChildBlock__unstable } from 'vtex.render-runtime'
+import LegacyHeader from './legacy'
+import CustomHeader from './CustomHeader'
 
-import styles from './store-header.css'
-import useDevice from './hooks/useDevice'
+const Header = props => {
+  const headerDesktop = useChildBlock__unstable({ id: 'header-desktop' })
+  const headerMobile = useChildBlock__unstable({ id: 'header-mobile' })
 
-/**
- * Main header component
- */
-const Header = ({
-  leanWhen,
-  linkUrl,
-  logoUrl,
-  logoTitle,
-  logoSize,
-  showSearchBar,
-  showLogin,
-  iconClasses,
-  labelClasses,
-  collapsibleAnimation,
-}) => {
-  const { page, getSettings } = useRuntime()
-  const { mobile } = useDevice()
-  const [containerHeight, setContainerHeight] = useState(null)
-  const { storeName } = getSettings('vtex.store')
+  const hasCustomHeader = headerDesktop || headerMobile
 
-  useEffect(() => {
-    if (document) {
-      const containerElement = document.querySelector(
-        '.vtex-store-header-2-x-container'
-      )
-      const elementHeight = containerElement && containerElement.offsetHeight
-      setContainerHeight(elementHeight)
-    }
+  console.log({
+    headerDesktop, headerMobile, hasCustomHeader
   })
 
-  const topMenuOptions = {
-    linkUrl,
-    logoUrl,
-    logoTitle: logoTitle || storeName || '',
-    logoSize,
-    showSearchBar,
-    showLogin,
-    iconClasses,
-    labelClasses,
-    collapsibleAnimation,
-    mobile,
-  }
-
-  const isLeanMode = () => {
-    const acceptedPaths = new RegExp(leanWhen)
-    return acceptedPaths.test(page)
-  }
-
-  const containerClasses = classNames(
-    styles.container,
-    'fixed top-0 z-4 w-100',
-    isLeanMode() ? styles.leanMode : ''
-  )
-
   return (
-    <Fragment>
-      <div className={containerClasses}>
-        <TopMenu
-          {...topMenuOptions}
-          leanMode={isLeanMode()}
-          extraHeaders={
-            <div
-              className="left-0 w-100"
-              style={{
-                transform: 'translateZ(0)', //Avoid shaking
-              }}
-            >
-              <ExtensionPoint id="telemarketing" />
-              <ExtensionPoint id="menu-link" />
-            </div>
-          }
-        />
-      </div>
-      <Spacer containerHeight={containerHeight} />
-    </Fragment>
+    hasCustomHeader 
+      ? <CustomHeader {...props} />
+      : <LegacyHeader {...props} />
   )
-}
-
-Header.propTypes = {
-  /** Cases in which the menu is in lean mode */
-  leanWhen: PropTypes.string,
-  ...login.propTypes,
-  ...searchBar.propTypes,
-  ...icons.propTypes,
-  ...logo.propTypes,
-  ...collapsible.propTypes,
-}
-
-Header.defaultProps = {
-  leanWhen: 'a^',
-  ...login.defaultProps,
-  ...searchBar.defaultProps,
-  ...icons.defaultProps,
-  ...logo.defaultProps,
-  ...collapsible.defaultProps,
-}
-
-Header.schema = {
-  title: 'editor.header.title',
-  description: 'editor.header.description',
-  type: 'object',
-  properties: {
-    logoUrl: {
-      type: 'string',
-      title: 'editor.header.logo.image',
-      widget: {
-        'ui:widget': 'image-uploader',
-      },
-    },
-    linkUrl: {
-      type: 'string',
-      title: 'editor.header.link.url',
-    },
-    showSearchBar: {
-      title: 'editor.header.show.searchbar.title',
-      type: 'boolean',
-      default: true,
-      isLayout: true,
-    },
-    showLogin: {
-      title: 'editor.header.show.login.title',
-      type: 'boolean',
-      default: true,
-      isLayout: true,
-    },
-  },
 }
 
 export default Header
