+++ react/legacy/__mocks__/vtex.store-components.js
