+++ react/legacy/__mocks__/vtex.store-icons.js
