+++ react/legacy/__mocks__/vtex.styleguide.js
