+++ react/legacy/components/Collapsible.js
