+++ react/legacy/components/Logo.js
