+++ react/legacy/components/SearchBar.js
