+++ react/legacy/defaults.js
