+++ react/legacy/hooks/useDevice.js
