+++ react/package.json
@@ -16,6 +16,7 @@
     "react-adopt": "^0.6.0",
     "react-dom": "^16.3.2",
     "react-intl": "^2.4.0",
+    "react-media": "^1.9.2",
     "react-resize-detector": "^3.0.1",
     "react-spring": "^8.0.7",
     "throttleit": "^1.0.0"
